+++ src/constants/apiPermissions.js
@@ -2,7 +2,7 @@ const roles = require('./roles')
 
 module.exports = {
   User: {
-    
+
   },
   Event: {
