+++ src/middleware/fieldPermissions.js
@@ -1,23 +1,7 @@
 const boom = require('boom')
 const { messages: errMessages } = require('../constants/errors.json')
-const { hasSufficientRole } = require('./permissions')
 const roles = require('../constants/roles.js')
-
-const orderedRoles = [roles.BASIC, roles.ADMIN, roles.SUPER]
-
-const getUnauthorizedFields = fieldPermissions => fieldsToChange => user => {
-  const permissionedFields = Object.keys(fieldPermissions)
-
-  return fieldsToChange
-    // filter out fields which are not permissioned
-    .filter(field => permissionedFields.includes(field))
-      // filter down to fields which user is not permitted to write to
-    .filter(field => {
-      const [ minRole, ownerIsPermitted ] = fieldPermissions[field]
-      return !hasSufficientRole({ minRole })(user) &&
-        !(ownerIsPermitted && user && user.isResourceOwner)
-    })
-}
+const { orderedRoles, getUnauthorizedFields } = require('../helpers/permissions')
 
 module.exports = fieldPermissions => {
   // fieldPermissions should be an object with the following form
@@ -29,7 +13,7 @@ module.exports = fieldPermissions => {
   // check supplied fieldPermissions are in correct form
   permissionedFields.forEach(field => {
     const [ minRole, owner ] = fieldPermissions[field]
-    const ownerIsPermitted = !!owner
+    // const ownerIsPermitted = !!owner
 
     if (!orderedRoles.includes(minRole)) {
       throw boom.badImplementation()
@@ -40,6 +24,7 @@ module.exports = fieldPermissions => {
   })
 
   return (req, res, next) => {
+    // const resourceType = permissions.getResourceType(req)
     const fieldsToChange = Object.keys(req.body)
 
     const unauthorizedFields =
@@ -53,5 +38,3 @@ module.exports = fieldPermissions => {
     next()
   }
 }
-
-module.exports.getUnauthorizedFields = getUnauthorizedFields
