+++ src/middleware/permissions.js
@@ -1,52 +1,12 @@
-const url = require('url')
 const boom = require('boom')
-const roles = require('../constants/roles.js')
-const { auth, messages: errMessages } = require('../constants/errors.json')
-const { rejectIfNull } = require('../db/utils')
-
-const Event = require('../models/Event')
-const Place = require('../models/Place')
-const Product = require('../models/Product')
-const User = require('../models/User')
-
-const orderedRoles = [roles.BASIC, roles.ADMIN, roles.SUPER]
-
-const hasSufficientRole = ({ minRole }) => user => {
-  // throw error on initialisation if passed option is not valid
-  if (!orderedRoles.includes(minRole)) {
-    throw boom.badImplementation()
-  }
-
-  return user && (orderedRoles.indexOf(user.role) >= orderedRoles.indexOf(minRole))
-}
-
-const checkUserOwnsResource = resourceType => resourceId => user => {
-  if (resourceType === User) {
-    return resourceId === user.id.toString()
-      ? Promise.resolve()
-      : Promise.reject(boom.unauthorized(auth.UNAUTHORIZED))
-  }
-
-  return resourceType.findById(resourceId)
-  .then(rejectIfNull(errMessages.GET_ID_NOT_FOUND))
-  // currently using toString(), may be better to use <mongoid>.equals(<mongoid>)
-  .then(doc => doc.owner.toString() === user.id.toString()
-    ? Promise.resolve()
-    : Promise.reject(boom.unauthorized(auth.UNAUTHORIZED))
-  )
-}
-
-const getResourceType = req => {
-  // currently needs url to be of the form '/<resource>/:id'
-  const routeResourceMapping = {
-    events: Event,
-    places: Place,
-    products: Product,
-    users: User
-  }
-  const pathName = url.parse(req.url).pathname
-  return routeResourceMapping[pathName.split('/')[1]]
-}
+const { auth } = require('../constants/errors.json')
+const roles = require('../constants/roles')
+const {
+  orderedRoles,
+  hasSufficientRole,
+  checkUserOwnsResource,
+  getResourceType
+} = require('../helpers/permissions')
 
 module.exports = ({ authorizedRoles }) => {
   // authorizedRoles should be an array [ minRole [, OWNER] ]
@@ -93,7 +53,3 @@ module.exports = ({ authorizedRoles }) => {
     .catch(err => next(err))
   }
 }
-
-// export functions for testing
-module.exports.hasSufficientRole = hasSufficientRole
-module.exports.checkUserOwnsResource = checkUserOwnsResource
